+++ client/src/App.jsx
@@ -17,6 +17,7 @@ import AlumniDirectory from './pages/AlumniDirectory';
 import ReferralBoard from './pages/ReferralBoard';
 import PublicProfile from './pages/PublicProfile';
 import ProfileSettings from './pages/ProfileSettings';
+import Reviews from './pages/Reviews';
 import ProtectedRoute from './components/ProtectedRoute';
 
 function App() {
@@ -37,6 +38,7 @@ function App() {
           <Route path="/announcements" element={<ProtectedRoute><Announcements /></ProtectedRoute>} />
           <Route path="/alumni" element={<ProtectedRoute><AlumniDirectory /></ProtectedRoute>} />
           <Route path="/referrals" element={<ProtectedRoute><ReferralBoard /></ProtectedRoute>} />
+          <Route path="/reviews" element={<ProtectedRoute><Reviews /></ProtectedRoute>} />
           <Route path="/moderation" element={<ProtectedRoute><ModerationQueue /></ProtectedRoute>} />
           <Route path="/chats" element={<ProtectedRoute><Chats /></ProtectedRoute>} />
           <Route path="/chat/:userId" element={<ProtectedRoute><ChatThread /></ProtectedRoute>} />