+++ client/src/components/Navbar.jsx
@@ -38,6 +38,7 @@ const Navbar = () => {
         <Link to="/announcements" style={{ color: 'white', textDecoration: 'none' }}>Announcements</Link>
         <Link to="/alumni" style={{ color: 'white', textDecoration: 'none' }}>Alumni</Link>
         <Link to="/referrals" style={{ color: 'white', textDecoration: 'none' }}>Referrals</Link>
+        <Link to="/reviews" style={{ color: 'white', textDecoration: 'none' }}>Reviews</Link>
         <Link to="/chats" style={{ color: 'white', textDecoration: 'none' }}>Inbox</Link>
         <NotificationsDropdown />
         <button